requirements.txt
Django==3.2
djangorestframework=3.13.1
markdown==3.4.1
django-filter==22.1
django-versatileimagefield==2.2
python-magic
django-model-utils